CityControllerIT - deleteShouldReturnNotFoundWhenNonExistingId

+++ src/main/java/com/devsuperior/demo/services/CityService.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.devsuperior.demo.dto.CityDTO;
 import com.devsuperior.demo.entities.City;
 import com.devsuperior.demo.repositories.CityRepository;
+import com.devsuperior.demo.services.exceptions.ResourceNotFoundException;
 
 
 
@@ -34,9 +35,9 @@ public class CityService {
 	}
 
 	public void delete(Long id) {
-//		if(!repository.existsById(id)) {
-//			throw new ResourceNotFoundException("Recurso não encontrado");			
-//		}
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");			
+		}
 		try {
 			repository.deleteById(id);
 		} catch (DataIntegrityViolationException e) {
